// EmbeddedReportsTemplate/client-app/app/shared/period-enum.ts
export enum PeriodEnum {
    January = 1,
    February = 2,
    March = 3,
    April = 4,
    May = 5,
    June = 6,
    July = 7,
    August = 8,
    September = 9,
    October = 10,
    November = 11,
    December = 12,
    FirstQuarter = 13,
    SecondQuarter = 14,
    ThirdQuarter = 15,
    FourthQuarter = 16,
    Yearly = 17,
}